Q2/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.12
numpy>=1.24.0
tqdm>=4.66.0
pillow>=10.0.0
tensorboard>=2.14.0
torchinfo>=1.8.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.11.0
